Move Label component to shared
